+++ projects/ngx-form-object/src/lib/form-object/form-object.ts
@@ -17,63 +17,63 @@ const defaultModelOptions: FormObjectOptions = {
 	getModelType: (model: any) => model.constructor.name,
 };
 
-export abstract class FormObject {
+export abstract class FormObject<T> {
 	public _options: FormObjectOptions;
-	public validators: object = {};
+	public validators: Record<string, ValidatorFn | Array<ValidatorFn>> = {};
 	public formGroupOptions: FormGroupOptions = {};
-	public formStoreClass: any;
+	public formStoreClass: new () => FormStore<T>;
 
-	protected beforeSave(store: FormStore): Observable<FormStore> {
+	protected beforeSave(store: FormStore<T>): Observable<FormStore<T>> {
 		return observableOf(store);
 	}
 
 	// @ts-ignore
-	protected save(model: any): Observable<any> {
+	protected save(model: T): Observable<T> {
 		return throwError('Save function must be implemented in the corresponding form object');
 	}
 
-	protected afterSave(model?: any, _form?: FormStore): Observable<any> {
+	protected afterSave(model?: T, _form?: FormStore<T>): Observable<T> {
 		return observableOf(model);
 	}
 
-	constructor(public model: any, protected options: FormObjectOptions) {
+	constructor(public model: T, protected options: FormObjectOptions) {
 		this._options = {
 			...defaultModelOptions,
 			...options,
 		};
 	}
 
-	get attributeProperties(): Map<string | symbol, PropertyOptions> {
+	public get attributeProperties(): Map<string | symbol, PropertyOptions> {
 		const modelMetadata: ModelMetadata =
 			Reflect.getMetadata(MetadataProperty.MODEL_METADATA, this.model.constructor) || {};
 		return modelMetadata.attributeProperties || new Map();
 	}
 
-	get attributePropertiesKeys(): Array<string | symbol> {
+	public get attributePropertiesKeys(): Array<string | symbol> {
 		return Array.from(this.attributeProperties.keys());
 	}
 
-	get hasManyProperties(): Map<string | symbol, PropertyOptions> {
+	public get hasManyProperties(): Map<string | symbol, PropertyOptions> {
 		const modelMetadata: ModelMetadata =
 			Reflect.getMetadata(MetadataProperty.MODEL_METADATA, this.model.constructor) || {};
 		return modelMetadata.hasManyProperties || new Map();
 	}
 
-	get hasManyPropertiesKeys(): Array<string | symbol> {
+	public get hasManyPropertiesKeys(): Array<string | symbol> {
 		return Array.from(this.hasManyProperties.keys());
 	}
 
-	get belongsToProperties(): Map<string | symbol, PropertyOptions> {
+	public get belongsToProperties(): Map<string | symbol, PropertyOptions> {
 		const modelMetadata: ModelMetadata =
 			Reflect.getMetadata(MetadataProperty.MODEL_METADATA, this.model.constructor) || {};
 		return modelMetadata.belongsToProperties || new Map();
 	}
 
-	get belongsToPropertiesKeys(): Array<string | symbol> {
+	public get belongsToPropertiesKeys(): Array<string | symbol> {
 		return Array.from(this.belongsToProperties.keys());
 	}
 
-	public getModelType(model: any): string {
+	public getModelType(model: T): string {
 		if (this._options.getConfig) {
 			// TODO see if can be removed
 			return this._options.getConfig(this.model.constructor).type;
@@ -86,21 +86,22 @@ export abstract class FormObject {
 		const validators = this.validators[attributeName];
 
 		if (validators && validators.length > 1) {
-			return Validators.compose(validators);
+			return Validators.compose(validators as Array<ValidatorFn>);
 		} else {
 			return validators;
 		}
 	}
 
-	public reset(form: any): void {
+	public reset(form: FormStore<T>): void {
 		this.rollbackAttributes(form);
 		this.rollbackBelongsToRelationships(form);
 		this.rollbackHasManyRelationships(form);
 	}
 
-	public mapPropertiesToModel(form: any): void {
-		this.attributePropertiesKeys.forEach((propertyName: string | symbol) => {
-			const formProperty = form.controls[propertyName];
+	// should be renamed to mapAttributePropertiesToModel
+	public mapPropertiesToModel(form: FormStore<T>): void {
+		this.attributePropertiesKeys.forEach((propertyName: string) => {
+			const formProperty = form.controls[propertyName] as ExtendedFormControl;
 
 			if (formProperty.isChanged) {
 				const unmaskFunction: Function = this[`unmask${capitalize(propertyName.toString())}`]; // tslint:disable-line: ban-types
@@ -114,9 +115,9 @@ export abstract class FormObject {
 		});
 	}
 
-	public mapBelongsToPropertiesToModel(form: any): void {
-		this.belongsToPropertiesKeys.forEach((propertyName) => {
-			const formProperty = form.controls[propertyName];
+	public mapBelongsToPropertiesToModel(form: FormStore<T>): void {
+		this.belongsToPropertiesKeys.forEach((propertyName: string) => {
+			const formProperty = form.controls[propertyName] as FormStore<unknown>;
 
 			if (formProperty.isChanged) {
 				if (formProperty.formObject) {
@@ -128,14 +129,14 @@ export abstract class FormObject {
 		});
 	}
 
-	public isFormValid(form: FormStore): boolean {
+	public isFormValid(form: FormStore<T>): boolean {
 		return form.valid || form.disabled;
 	}
 
-	public _save(form: FormStore): Observable<any> {
+	public _save(form: FormStore<T>): Observable<T> {
 		return observableOf(true).pipe(
 			flatMap(() => this._beforeSave(form)),
-			flatMap((validFormStore: FormStore) => {
+			flatMap((validFormStore: FormStore<T>) => {
 				const validatedFormWithModel$ = new ReplaySubject();
 
 				this.save(this.model)
@@ -145,7 +146,7 @@ export abstract class FormObject {
 							return throwError(error);
 						})
 					)
-					.subscribe((savedModel: any) => {
+					.subscribe((savedModel: T) => {
 						validatedFormWithModel$.next({
 							savedModel,
 							validFormStore,
@@ -160,9 +161,9 @@ export abstract class FormObject {
 		);
 	}
 
-	protected rollbackAttributes(form: any): void {
-		this.attributePropertiesKeys.forEach((propertyName: string | symbol) => {
-			const formProperty = form.controls[propertyName];
+	protected rollbackAttributes(form: FormStore<T>): void {
+		this.attributePropertiesKeys.forEach((propertyName: string) => {
+			const formProperty = form.controls[propertyName] as ExtendedFormControl;
 			const unmaskFunction: Function = this[`mask${capitalize(propertyName.toString())}`]; // tslint:disable-line: ban-types
 
 			const propertyValue: any = unmaskFunction
@@ -175,9 +176,9 @@ export abstract class FormObject {
 		});
 	}
 
-	protected rollbackBelongsToRelationships(form: any): void {
-		this.belongsToPropertiesKeys.forEach((propertyName) => {
-			const formProperty = form.controls[propertyName];
+	protected rollbackBelongsToRelationships(form: FormStore<T>): void {
+		this.belongsToPropertiesKeys.forEach((propertyName: string) => {
+			const formProperty = form.controls[propertyName] as ExtendedFormControl;
 
 			if (formProperty.isChanged) {
 				form.controls[propertyName].setValue(this.model[propertyName]);
@@ -185,8 +186,8 @@ export abstract class FormObject {
 		});
 	}
 
-	protected rollbackHasManyRelationships(form: any): void {
-		this.hasManyPropertiesKeys.forEach((propertyName) => {
+	protected rollbackHasManyRelationships(form: FormStore<T>): void {
+		this.hasManyPropertiesKeys.forEach((propertyName: string) => {
 			const formProperty = form.controls[propertyName];
 
 			const rollback: Function = this[`rollback${capitalize(propertyName.toString())}`]; // tslint:disable-line: ban-types
@@ -197,9 +198,9 @@ export abstract class FormObject {
 		});
 	}
 
-	private _beforeSave(form: FormStore): Observable<FormStore> {
-		const form$: Observable<FormStore> = this.beforeSave(form).pipe(
-			flatMap((transformedForm: FormStore) => {
+	private _beforeSave(form: FormStore<T>): Observable<FormStore<T>> {
+		const form$: Observable<FormStore<T>> = this.beforeSave(form).pipe(
+			flatMap((transformedForm: FormStore<T>) => {
 				this.mapPropertiesToModel(transformedForm);
 				this.mapBelongsToPropertiesToModel(transformedForm);
 
@@ -214,9 +215,9 @@ export abstract class FormObject {
 		return form$;
 	}
 
-	private _afterSave(model: any, form: FormStore): Observable<any> {
-		const form$: Observable<any> = this.afterSave(model, form).pipe(
-			flatMap((transformedModel: any) => {
+	private _afterSave(model: T, form: FormStore<T>): Observable<T> {
+		const form$: Observable<T> = this.afterSave(model, form).pipe(
+			flatMap((transformedModel: T) => {
 				this.mapModelPropertiesToForm(transformedModel, form);
 				this.resetBelongsToFormControls(transformedModel, form);
 				return observableOf(transformedModel);
@@ -226,7 +227,7 @@ export abstract class FormObject {
 		return form$;
 	}
 
-	private mapModelPropertiesToForm(model: any, form: FormStore): void {
+	private mapModelPropertiesToForm(model: T, form: FormStore<T>): void {
 		this.attributePropertiesKeys.forEach((propertyName: string) => {
 			const formControl: ExtendedFormControl = form.controls[propertyName] as ExtendedFormControl;
 
@@ -239,7 +240,7 @@ export abstract class FormObject {
 		});
 	}
 
-	private resetBelongsToFormControls(_model: any, form: FormStore): void {
+	private resetBelongsToFormControls(_model: T, form: FormStore<T>): void {
 		this.belongsToPropertiesKeys.forEach((propertyName: string) => {
 			const formControl: ExtendedFormControl = form.controls[propertyName] as ExtendedFormControl;
 			if (formControl.resetValue) {