+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 8.0.0-beta
+
+  * [BREAKING CHANGE] `FormStore` uses a generic now
+  * [BREAKING CHANGE] `FormObject` uses a generic now
+  * [BREAKING CHANGE] `FormObjectBuilder` uses a generic now
+
 ### 9.0.3-beta
 
   * Temporary adjust save method types